Implement q_delete_mid function
The q_delete_mid function repeatedly moves one
pointer from the head and the other one from the tail until they meet
each other. At this point, the pointer named right is pointing to the
middle node.

Once the middle node is located, it is removed and its memory
is freed with q_release_element function.

The function ensures that an empty list or a NULL head is handled
safely by returning false.

Change-Id: Id70b14e3f886e775ac01e3411d095937f7ab93e2